style: 添加布局优化样式
