+++ apps/sidebar-app/src/App.module.scss
@@ -1,3 +1,6 @@
 .app {
-  display: flex;
+  position: relative;
+  .main {
+    margin-left: 80px;
+  }
 }
