+++ apps/sidebar-app/src/components/RouterConfig.tsx
@@ -5,13 +5,19 @@ import ExternalUser from "../pages/ExternalUser";
 import ExternalChat from "../pages/ExternalChat";
 import Action from "../pages/Action";
 
+export const routes = [
+  {url: '/', label: '首页', key: 'home', page: Home},
+  {url: '/external-user', label: '私聊', key: 'external-user', page: ExternalUser},
+  {url: '/external-chat', label: '群聊', key: 'external-chat', page: ExternalChat},
+  {url: '/action', label: '操作', key: 'action', page: Action},
+]
+
 const RouterConfig: FC = () => {
   return (
     <Routes>
-      <Route path="/" element={<Home/>}/>
-      <Route path="/external-user" element={<ExternalUser/>}/>
-      <Route path="/external-chat" element={<ExternalChat/>}/>
-      <Route path="/action" element={<Action/>}/>
+      {routes.map(route => (
+        <Route key={route.key} path={route.url} element={<route.page/>}/>
+      ))}
     </Routes>
   )
 }
