+++ apps/sidebar-app/src/components/VerticalMenu/styles.module.scss
@@ -1,0 +1,9 @@
+.menu {
+  position: absolute;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  .item {
+    text-align: center;
+  }
+}
